requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.4
sqlalchemy==2.0.25
openpyxl==3.1.2
python-dotenv==1.0.0
typing-extensions==4.9.0
